Mini bir sınıf/nesne örneği eklenir

+++ php/mysql-articles/delete.php
@@ -1,20 +1,24 @@
 <?php
 
 require_once "init.php";
+if (!isset($_SESSION['is_admin'])) {
+    header("Location: index.php");
+    die();
+}
 
-if(isset($_GET['id'])){
-  $article = $connection->prepare("DELETE FROM articles WHERE id = ?");
-  $article->execute(array($_GET['id']));
-  $result = $article->rowCount();
-}else{
-  $result = false;
+if (isset($_GET['id'])) {
+    $article = $connection->prepare("DELETE FROM articles WHERE id = ?");
+    $article->execute(array($_GET['id']));
+    $result = $article->rowCount();
+} else {
+    $result = false;
 }
 ?>
 
 <h3>
-  <?php if($result): ?>
+  <?php if ($result): ?>
     Kayıt başarıyla silindi.
-  <?php else:?>
+  <?php else: ?>
     HATA: İşlem tamamlanamadı.
   <?php endif;?>
 </h3>
